+++ authpass/lib/utils/logging_utils.dart
@@ -2,6 +2,8 @@ import 'dart:io';
 import 'dart:isolate';
 
 import 'package:authpass/utils/path_utils.dart';
+import 'package:authpass/utils/platform.dart';
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:logging/logging.dart';
 import 'package:logging_appenders/logging_appenders.dart';
@@ -62,6 +64,78 @@ class LoggingUtils {
     });
     Isolate.current.addOnExitListener(exitPort.sendPort, response: 'exit');
   }
+
+  static Future<Map<String, dynamic>> getDebugDeviceInfo() async {
+    final di = DeviceInfoPlugin();
+    if (AuthPassPlatform.isWeb) {
+      return (await di.webBrowserInfo).importantInfo();
+    }
+    if (AuthPassPlatform.isAndroid) {
+      return (await di.androidInfo).importantInfo();
+    }
+    if (AuthPassPlatform.isIOS) {
+      return (await di.iosInfo).importantInfo();
+    }
+    if (AuthPassPlatform.isLinux) {
+      return (await di.linuxInfo).importantInfo();
+    }
+    if (AuthPassPlatform.isWindows) {
+      return (await di.windowsInfo).importantInfo();
+    }
+    if (AuthPassPlatform.isMacOS) {
+      return (await di.macOsInfo).importantInfo();
+    }
+    return <String, dynamic>{
+      'unknownPlatform': AuthPassPlatform.operatingSystem
+    };
+  }
+}
+
+extension on AndroidDeviceInfo {
+  Map<String, dynamic> importantInfo() => <String, dynamic>{
+        'board': board,
+        'device': device,
+        'hardware': hardware,
+        'manufacturer': manufacturer,
+        'product': product,
+        'version': version,
+      };
+}
+
+extension on IosDeviceInfo {
+  Map<String, dynamic> importantInfo() => <String, dynamic>{
+        'localizedModel': localizedModel,
+        'mode': model,
+        'name': name,
+      };
+}
+
+extension on WebBrowserInfo {
+  Map<String, dynamic> importantInfo() => <String, dynamic>{
+        'userAgent': userAgent,
+      };
+}
+
+extension on LinuxDeviceInfo {
+  Map<String, dynamic> importantInfo() => <String, dynamic>{
+        'name': name,
+        'version': version,
+      };
+}
+
+extension on WindowsDeviceInfo {
+  Map<String, dynamic> importantInfo() => <String, dynamic>{
+        'systemMemoryInMegabytes': systemMemoryInMegabytes,
+      };
+}
+
+extension on MacOsDeviceInfo {
+  Map<String, dynamic> importantInfo() => <String, dynamic>{
+        'arch': arch,
+        'model': model,
+        'hostName': hostName,
+        'osRelease': osRelease,
+      };
 }
 
 class StringBufferWrapper with ChangeNotifier {