+++ authpass/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `CONTRIBUTORS.md` to about dialog. #237 (@ATofighi)
 * Fixed potential crash in Preferences screen.
 * Fixed broken 'manage groups'.
+* Add links to forum, report error to forum https://forum.authpass.app/
 
 # 1.8.1 - 2021-07-13